Refactor to make genomicsdb_query and genomicsdb_cache console scripts
